LayoutTests/editing/pasteboard/paste-before-tab-span-expected.txt

This test ensures pasting content before a tab span isn't pasted into the tab span.

Result:
| <span>
|   id="test"
|   "hellohelloa<#selection-caret>"
|   <span>
|     class="Apple-tab-span"
|     style="white-space:pre"
|     "		"
|   "world"
| "
"
